+++ src/client-e2e-test/runCursorlessCommand.ts
@@ -1,0 +1,7 @@
+import * as vscode from "vscode";
+import { cursorlessCommandId } from "../common/commandIds";
+import { Command } from "../core/commandRunner/command.types";
+
+export function runCursorlessCommand(command: Command) {
+  return vscode.commands.executeCommand(cursorlessCommandId, command);
+}
