+++ src/common/commandIds.ts
@@ -1,0 +1,1 @@
+export const cursorlessCommandId = "cursorless.command";
