+++ src/core/commandRunner/CommandRunner.ts
@@ -1,6 +1,8 @@
 import * as vscode from "vscode";
 import { ActionType } from "../../actions/actions.types";
+import { cursorlessCommandId } from "../../common/commandIds";
 import { OutdatedExtensionError } from "../../errors";
+import { getActiveTextEditor } from "../../ide/activeTextEditor";
 import processTargets from "../../processTargets";
 import isTesting from "../../testUtil/isTesting";
 import { Target } from "../../typings/target.types";
@@ -19,9 +21,6 @@ import inferFullTargets from "../inferFullTargets";
 import { ThatMark } from "../ThatMark";
 import { Command } from "./command.types";
 import { selectionToThatTarget } from "./selectionToThatTarget";
-import { getActiveTextEditor } from "../../ide/activeTextEditor";
-
-const cursorlessCommand = "cursorless.command";
 
 // TODO: Do this using the graph once we migrate its dependencies onto the graph
 export default class CommandRunner {
@@ -39,7 +38,7 @@ export default class CommandRunner {
 
     this.disposables.push(
       vscode.commands.registerCommand(
-        cursorlessCommand,
+        cursorlessCommandId,
         this.runCommandBackwardCompatible,
       ),
     );
@@ -250,8 +249,4 @@ function constructThatTarget(
   } else {
     return targets;
   }
-}
-
-export function runCursorlessCommand(command: Command) {
-  return vscode.commands.executeCommand(cursorlessCommand, command);
 }
