+++ src/test/suite/groupByDocument.test.ts
@@ -1,6 +1,6 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
-import { runCursorlessCommand } from "../../core/commandRunner/CommandRunner";
+import { runCursorlessCommand } from "../../client-e2e-test/runCursorlessCommand";
 import HatTokenMap from "../../core/HatTokenMap";
 import { getCursorlessApi } from "../../util/getExtensionApi";
 import { standardSuiteSetup } from "./standardSuiteSetup";
